backend/requirements.txt
Django>=4.2,<5.0
django-modeltranslation>=0.18
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.0
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.0
